added terminal command piping 	can run more than one program in a command an chain them renamed goto to dir due to goto being a groovy reserved word
