+++ core/assets/home/programs/print/systems/BootSystem.groovy
@@ -10,12 +10,12 @@ import os.systems.*;
 
 public class BootSystem extends EntitySystem {	
 
+    ProgramOutputComponent outputComponent = null;
+    ProgramInputComponent inputComponent = null;
+    
     public void addedToEngine(Engine engine) {
         ImmutableArray<Entity> inject = engine.getEntitiesFor(Family.all(ProgramEntityInjectionComponent.class).get());
         ProgramEntityInjectionComponent injection = inject.first().getComponent(ProgramEntityInjectionComponent.class);
-        
-        ProgramOutputComponent outputComponent = null;
-        ProgramInputComponent inputComponent = null;
         
         for(Entity e : injection.entities) {
             engine.addEntity(e);  
@@ -27,14 +27,13 @@ public class BootSystem extends EntitySystem {
             }
         }
         
-        if(outputComponent != null && inputComponent != null && inputComponent.input.size() > 0) {
-            inputComponent.input.poll();
-            outputComponent.output.add(inputComponent.input.poll());
-        }
+        if(inputComponent != null) inputComponent.input.poll();
     }
     
     public void update(float delta) {
-        
+        if(outputComponent != null && inputComponent != null && inputComponent.input.size() > 0) {
+            outputComponent.output.add(inputComponent.input.poll());
+        }
     }
     
     public void removedFromEngine(Engine engine) {
