+++ core/assets/home/programs/terminal/components/PipedCommandComponent.groovy
@@ -1,0 +1,7 @@
+package terminal.components;
+
+import com.badlogic.ashley.core.*;
+
+public class PipedCommandComponent implements Component {
+}
+
