+++ core/assets/home/programs/terminal/systems/RunCommandOnKeyEnterSystem.groovy
@@ -1,17 +1,16 @@
 package terminal.systems;
 
+import java.util.Arrays;
+
 import com.janfic.useride.kernel.components.*;
 import com.janfic.useride.kernel.systems.*;
 import com.badlogic.ashley.core.*;
 import com.badlogic.ashley.utils.*;
 import com.badlogic.gdx.utils.viewport.*;
-import com.badlogic.gdx.graphics.g2d.*;
-import com.badlogic.gdx.graphics.*;
 import com.badlogic.gdx.utils.*;
 import com.badlogic.gdx.assets.*;
 import com.badlogic.gdx.*;
 import com.badlogic.gdx.input.*;
-import com.badlogic.gdx.math.*;
 import com.badlogic.gdx.files.FileHandle;
 import terminal.components.*;
 import os.components.*;
@@ -50,42 +49,51 @@ public class RunCommandOnKeyEnterSystem extends EntitySystem {
             FileComponent dir = fileMapper.get(entity);
             
             if(keyInput.keyTyped == 13 && command.text.length() > 0) {
-                String[] text = command.text.split(" ");
-                if(text.length > 0) {
-                    String program = text[0];
-                    String input = null;
-                    if(text.length > 1) {
-                        input = command.text.substring(command.text.indexOf(program) + program.length() + 1);
-                    }
+                String[] commands = command.text.split("\\s*\\|\\s*");
+                System.out.println(Arrays.toString(commands));
+                ProgramOutputComponent previousOutput = null;
+                for(int i = 0; i < commands.length; i++) {
+                    String[] text = commands[i].split(" ");
+                    System.out.println(Arrays.toString(text));
+                    if(text.length > 0) {
+                        String program = text[0].trim();
+                        String input = null;
+                        if(text.length > 1) {
+                            input = commands[i].substring(commands[i].indexOf(program) + program.length() + 1);
+                        }
                     
-                    FileHandle programFolder = Gdx.files.local("home/programs/" + program);
+                        FileHandle programFolder = Gdx.files.local("home/programs/" + program);
                     
-                    if(programFolder.exists()) {
-                        Entity injectionEntity = new Entity();
+                        if(programFolder.exists()) {
+                            Entity injectionEntity = new Entity();
                         
-                        ProgramInputComponent programInput = new ProgramInputComponent();
-                        programInput.input.add(dir);
-                        if(input != null) programInput.input.add(input);
-                        ProgramOutputComponent programOutput = new ProgramOutputComponent();
-                    
-                        injectionEntity.add(programInput);
-                        injectionEntity.add(programOutput);
-
+                            ProgramInputComponent programInput = new ProgramInputComponent();
+                            if(previousOutput != null) programInput.input = previousOutput.output;
+                            programInput.input.add(dir);
+                            if(input != null) programInput.input.add(input);
+                            
+                            ProgramOutputComponent programOutput = new ProgramOutputComponent();
+                            previousOutput = programOutput;
+                            
+                            injectionEntity.add(programInput);
+                            injectionEntity.add(programOutput);
+                            
+                            Entity runCommandAttempt = new Entity();
                         
-                        Entity runCommandAttempt = new Entity();
+                            ProgramStartRequestComponent startRequest = new ProgramStartRequestComponent(name: program);
+                            ProgramEntityInjectionComponent injection = new ProgramEntityInjectionComponent();
+                            injection.entities.add(injectionEntity);
+                            FileComponent fileComponent = new FileComponent(file: programFolder);
                         
-                        ProgramStartRequestComponent startRequest = new ProgramStartRequestComponent(name: program);
-                        ProgramEntityInjectionComponent injection = new ProgramEntityInjectionComponent();
-                        injection.entities.add(injectionEntity);
-                        FileComponent fileComponent = new FileComponent(file: programFolder);
-                        
-                        runCommandAttempt.add(startRequest);
-                        runCommandAttempt.add(injection);
-                        runCommandAttempt.add(fileComponent);
-                        runCommandAttempt.add(programOutput);
-                        runCommandAttempt.add(programInput);
-                     
-                        this.getEngine().addEntity(runCommandAttempt);
+                            runCommandAttempt.add(startRequest);
+                            runCommandAttempt.add(injection);
+                            runCommandAttempt.add(fileComponent);
+                            runCommandAttempt.add(programOutput);
+                            runCommandAttempt.add(programInput);
+                            if(i != commands.length - 1) runCommandAttempt.add(new PipedCommandComponent());
+                            
+                            this.getEngine().addEntity(runCommandAttempt);
+                        }
                     }
                 }
             }
